Remove debugging
