+++ firmware/inc/doz_clock.h
@@ -6,7 +6,6 @@
 #include "gps.h"
 #include "rtc.h"
 
-#define DEBUG   0
 
 typedef struct doz_clock_t
 {
