+++ firmware/src/event_queue.c
@@ -1,156 +1,180 @@
 #include "event_queue.h"
 
 #include <stdio.h>
-#include <string.h>
-#include <stdlib.h>
-#include <stdbool.h>
 
-static BtnEventType BUTTON_EVENT_ARRAY[NUM_PRESS_TYPES][NUM_BUTTON_ID];
-static EventQ *head;
+static struct {
+    EventId buffer[BUFFER_SIZE];
+    uint8_t head;
+    uint8_t tail;
+    bool is_empty;
+}ring_buff; // Ring buffer data structure
 
-static void bufferPush(Event event);
-static void bufferPop();
-static bool bufferFull();
-static bool bufferEmpty();
+static void bufferPush(EventId event);
+static EventId bufferPop(void);
+static bool bufferFull(void);
+static bool bufferEmpty(void);
 
-ClockStatus EQ_Init()
+/*
+ *  Public functions
+*/
+ClockStatus EventQ_Init()
 {
-    BtnEventType btnEventType;
-    for (unsigned pressType = 0; pressType < NUM_PRESS_TYPES; ++pressType) {
-        for (unsigned btnId = 0; btnId < NUM_BUTTON_ID; ++btnId) {
-            btnEventType.e_btn_id = btnId;
-            btnEventType.e_btn_press_type = pressType;
-            BUTTON_EVENT_ARRAY[pressType][btnId] = btnEventType;
-        }
-    }
-
-    head = NULL;
-
-    return CLOCK_OK;
+    // Reset ring buffer
+    ring_buff.head = 0;
+    ring_buff.tail = 0;
+    ring_buff.is_empty = 1;
 }
 
-ClockStatus EQ_GetEvent(Event *event)
+ClockStatus EventQ_GetEvent(EventId *event)
 {
-    if (bufferEmpty()) {
+    if (bufferEmpty())
+    {
+        *event = E_NONE;
         return CLOCK_FAIL;
     }
-
-    memcpy(event, &head->s_event, sizeof(*event));
-    bufferPop();
-
+    // Remove event from buffer
+    *event = bufferPop();
     return CLOCK_OK;
 }
 
-ClockStatus EQ_TriggerButtonEvent(BtnId button, BtnPressType type)
+ClockStatus EventQ_TriggerButtonEvent(BtnId button, BtnPressType type)
 {
-    Event event = { .e_event_type = BTN_EVENT,
-                    .s_btn_event = BUTTON_EVENT_ARRAY[type][button],
-                    .e_light_event = 0,
-                    .e_alarm_event = 0 };
-    bufferPush(event);
+    EventId btn_event = E_NONE;
+    if (bufferFull()) 
+    {
+        return CLOCK_FAIL;
+    }
+    switch (button)
+    {
+        case DISPLAY:
+            btn_event = (type == LONG) ? E_DISPLAY_LONG : E_DISPLAY_SHORT;
+            break;
+        case ALARM:
+            btn_event = (type == LONG) ? E_ALARM_LONG : E_ALARM_SHORT;
+            break;
+        case TIMER:
+            btn_event = (type == LONG) ? E_TIMER_LONG : E_TIMER_SHORT;
+            break;
+        case LEFT:
+            btn_event = (type == LONG) ? E_LEFT_LONG : E_LEFT_SHORT;
+            break;
+        case RIGHT:
+            btn_event = (type == LONG) ? E_RIGHT_LONG : E_RIGHT_SHORT;
+            break;
+        case UP:
+            btn_event = (type == LONG) ? E_UP_LONG : E_UP_SHORT;
+            break;
+        case DOWN:
+            btn_event = (type == LONG) ? E_DOWN_LONG : E_DOWN_SHORT;
+            break;
+        case DOZ:
+            btn_event = (type == LONG) ? E_DOZ_LONG : E_DOZ_SHORT;
+            break;
+        case TRAD:
+            btn_event = (type == LONG) ? E_TRAD_LONG : E_TRAD_SHORT;
+            break;
+        case VOLUP:
+            btn_event = (type == LONG) ? E_VOLUP_LONG : E_VOLUP_SHORT;
+            break;
+        case VOLDOWN:
+            btn_event = (type == LONG) ? E_VOLDOWN_LONG : E_VOLDOWN_SHORT;
+            break;
+        case CANCEL:
+            btn_event = (type == LONG) ? E_CANCEL_LONG : E_CANCEL_SHORT;
+            break;
+        default:
+            break;
+    }
+    if (btn_event == E_NONE)
+    {
+        return CLOCK_FAIL;
+    }
+    else
+    {
+        // Add event to buffer
+        bufferPush(btn_event);
+        return CLOCK_OK;
+    }
+}
+
+ClockStatus EventQ_TriggerLightEvent(LightEventType type)
+{
+    if (bufferFull()) 
+    {
+        return CLOCK_FAIL;
+    }
+    switch (type)
+    {
+        // Add event to buffer
+        case LIGHT_ROOM:
+            bufferPush(E_ROOM_LIGHT);
+            break;
+        case DARK_ROOM:
+            bufferPush(E_ROOM_DARK);
+            break;
+        default:
+            return CLOCK_FAIL;
+    }
     return CLOCK_OK;
 }
 
-ClockStatus EQ_TriggerLightEvent(LightEventType type)
+ClockStatus EventQ_TriggerAlarmEvent(AlarmEventType type)
 {
-    Event event = { .e_event_type = LIGHT_EVENT,
-                    .s_btn_event = { 0 },
-                    .e_light_event = type,
-                    .e_alarm_event = 0 };
-    bufferPush(event);
+    if (bufferFull()) 
+    {
+        return CLOCK_FAIL;
+    }
+    switch (type)
+    {
+        // Add event to buffer
+        case ALARM_TRIG:
+            bufferPush(E_ALARM_TRIG);
+            break;
+        case TIMER_TRIG:
+            bufferPush(E_TIMER_TRIG);
+            break;
+        default:
+            return CLOCK_FAIL;
+    }
     return CLOCK_OK;
 }
 
-ClockStatus EQ_TriggerAlarmEvent(AlarmEventType type)
+/*
+ *  Private functions
+*/
+void bufferPush(EventId event)
 {
-    Event event = { .e_event_type = ALARM_EVENT,
-                    .s_btn_event = { 0 },
-                    .e_light_event = 0,
-                    .e_alarm_event = type };
-    bufferPush(event);
-    return CLOCK_OK;
+    if (!bufferFull())
+    {
+        // Add event to tail and increment tail
+        ring_buff.buffer[ring_buff.tail] = event;
+        ring_buff.is_empty = 0;
+        ring_buff.tail = (ring_buff.tail + 1) % BUFFER_SIZE;
+    }
 }
 
-static void bufferPush(Event event)
+EventId bufferPop(void)
 {
-    if (bufferFull()) {
-        return;
+    EventId event = E_NONE;
+    if (!bufferEmpty())
+    {
+        // Take event from head and increment head
+        event = ring_buff.buffer[ring_buff.head];
+        ring_buff.head = (ring_buff.head + 1) % BUFFER_SIZE;
+        if(ring_buff.head == ring_buff.tail)
+        {
+            ring_buff.is_empty = 1;
+        }
     }
-
-    EventQ *entry = malloc(sizeof(*entry));
-    entry->s_event = event;
-    entry->next = NULL;
-
-    if (head == NULL) {
-        head = entry;
-        return;
-    }
-
-    EventQ *temp = head;
-    while (temp->next != NULL) {
-        temp = temp->next;
-    }
-    temp->next = entry;
+    return event;
 }
 
-static void bufferPop()
+bool bufferFull(void)
 {
-    EventQ *temp = head;
-    if (head->next != NULL) {
-        head = head->next;
-    } else {
-        head = NULL;
-    }
-    free(temp);
+    return !ring_buff.is_empty && (ring_buff.head == ring_buff.tail);
 }
 
-static bool bufferFull()
+bool bufferEmpty(void)
 {
-    if (bufferEmpty()) {
-        return false;
-    }
-
-    unsigned buffCount = 1;
-    EventQ *temp = head;
-    while (temp->next != NULL) {
-        temp = temp->next;
-        ++buffCount;
-    }
-
-    return (buffCount >= BUFFER_SIZE);
+    return ring_buff.is_empty && (ring_buff.head == ring_buff.tail);
 }
-
-static bool bufferEmpty()
-{
-    return (head == NULL);
-}
-
-
-#if DEBUG
-void printEventQ()
-{
-    printf("--------------- EQ PRINT START\n\n");
-    if (head == NULL) {
-        printf("Buffer is Empty\n");
-        printf("\n--------------- EQ PRINT END\n\n");
-        return;
-    }
-
-    EventQ *temp = head;
-    unsigned i = 1;
-    do {
-        printf("\nEvent %u -----------\n", i++);
-        printf("Event Type: %d\nBtn Event: %d %d\n",
-                temp->s_event.e_event_type,
-                temp->s_event.s_btn_event.e_btn_id,
-                temp->s_event.s_btn_event.e_btn_press_type);
-        printf("Light Event: %d\nAlarm Event: %d\n",
-                temp->s_event.e_light_event,
-                temp->s_event.e_alarm_event);
-
-        temp = temp->next;
-    } while (temp != NULL);
-
-    printf("\n--------------- EQ PRINT END\n\n");
-}
-#endif
